[CinnamonBurnMyWindows@klangman] v0.9.6 enable "Fire" effect
- Enabled the "Fire" effect
- Apply a visual fix to "Incinerate" that was fixed in the Gnome extension recently
- Reordered the code to make it a bit easier to maintain
