+++ CinnamonBurnMyWindows@klangman/files/CinnamonBurnMyWindows@klangman/effects/BrokenGlass.js
@@ -14,16 +14,18 @@
 
 'use strict';
 
+//import * as utils from '../utils.js';
+
 // We import some modules only in the Shell process as they are not available in the
 // preferences process. They are used only in the creator function of the ShaderFactory
 // which is only called within GNOME Shell's process.
-const {ShaderFactory} = require('./ShaderFactory.js');
+const {ShaderFactory} =  require('./ShaderFactory.js');
 const Clutter = imports.gi.Clutter;
 const GdkPixbuf = imports.gi.GdkPixbuf;
 const Cogl = imports.gi.Cogl;
-
 const Gettext = imports.gettext;
 const GLib = imports.gi.GLib;
+
 const UUID = "CinnamonBurnMyWindows@klangman";
 
 Gettext.bindtextdomain(UUID, GLib.get_home_dir() + "/.local/share/locale");
@@ -36,14 +38,21 @@ function _(text) {
   return locText;
 }
 
+
 //////////////////////////////////////////////////////////////////////////////////////////
-// This effects dissolves your windows into a cloud of dust. For this, it uses an       //
-// approach similar to the Broken Glass effect. A dust texture is used to segment the   //
-// window texture into a set of layers. Each layer is then moved, sheared and scaled    //
-// randomly. Just a few layers are sufficient to create the illusion of many individual //
-// dust particles.                                                                      //
-// This effect is not available on GNOME 3.3x, due to the limitation described in the   //
-// documentation of vfunc_paint_target further down in this file.                       //
+// This effect shatters the window into pieces. For an explanation how this works, look //
+// further down in this file into the documentation of the Shader class. This effect is //
+// not available on GNOME 3.3x, due to the limitation described in the documentation    //
+// of vfunc_paint_target further down in this file.                                     //
+// This shader creates a complex-looking effect with rather simple means. Here is how   //
+// it works: The window is drawn five times on top of each other (see the SHARD_LAYERS  //
+// constant in the GLSL code). Each layer only draws some of the shards, all layers     //
+// combined make up the entire window. The layers are then scaled, rotated, and moved   //
+// independently from each other - this creates the impression that all shards are      //
+// moving independently. In reality, there are only five groups of shards! Which shard  //
+// belongs to which layer is defined by the green channel of the texture                //
+// resources/img/shards.png. The red channel of the texture contains the distance to    //
+// the shard edges. This information is used to fade out the shards.                    //
 //////////////////////////////////////////////////////////////////////////////////////////
 
 // The effect class can be used to get some metadata (like the effect's name or supported
@@ -58,33 +67,51 @@ var Effect = class Effect {
   constructor() {
     this.shaderFactory = new ShaderFactory(Effect.getNick(), (shader) => {
       // Create the texture in the first call.
-      if (!this._dustTexture) {
-        const dustData    = GdkPixbuf.Pixbuf.new_from_file( GLib.get_home_dir() +
-           '/.local/share/cinnamon/extensions/' + UUID + '/resources/img/dust.png');
-        this._dustTexture = new Clutter.Image();
-        this._dustTexture.set_data(dustData.get_pixels(), Cogl.PixelFormat.RGB_888,
-                                   dustData.width, dustData.height, dustData.rowstride);
+      if (!this._shardTexture) {
+        const shardData    = GdkPixbuf.Pixbuf.new_from_file( GLib.get_home_dir() +
+           '/.local/share/cinnamon/extensions/' + UUID + '/resources/img/shards.png');
+        this._shardTexture = new Clutter.Image();
+        this._shardTexture.set_data(shardData.get_pixels(), Cogl.PixelFormat.RGB_888,
+                                    shardData.width, shardData.height,
+                                    shardData.rowstride);
       }
 
-      // Store uniform locations of newly created shaders.
-      shader._uDustTexture = shader.get_uniform_location('uDustTexture');
-      shader._uDustColor   = shader.get_uniform_location('uDustColor');
-      shader._uSeed        = shader.get_uniform_location('uSeed');
-      shader._uDustScale   = shader.get_uniform_location('uDustScale');
+      // Store all uniform locations.
+      shader._uShardTexture = shader.get_uniform_location('uShardTexture');
+      shader._uSeed         = shader.get_uniform_location('uSeed');
+      shader._uEpicenter    = shader.get_uniform_location('uEpicenter');
+      shader._uShardScale   = shader.get_uniform_location('uShardScale');
+      shader._uBlowForce    = shader.get_uniform_location('uBlowForce');
+      shader._uGravity      = shader.get_uniform_location('uGravity');
 
-      // Write all uniform values at the start of each animation.
-      shader.connect('begin-animation', (shader, settings, forOpening, testMode) => {
-        // The dust particles will fade to this color over time.
-        const c = Clutter.Color.from_string(settings.get_string('snap-color'))[1];
+      // And update all uniforms at the start of each animation.
+      shader.connect(
+        'begin-animation', (shader, settings, forOpening, testMode, actor) => {
+          // Usually, the shards fly away from the center of the window.
+          let epicenterX = 0.5;
+          let epicenterY = 0.5;
 
-        // clang-format off
-        shader.set_uniform_float(shader._uDustColor, 4, [c.red / 255, c.green / 255, c.blue / 255, c.alpha / 255]);
-        shader.set_uniform_float(shader._uSeed,      2, [testMode ? 0 : Math.random(), testMode ? 0 : Math.random()]);
-        shader.set_uniform_float(shader._uDustScale, 1, [settings.get_double('snap-scale')]);
-        // clang-format on
-      });
+          // However, if this option is set, we use the mouse pointer position.
+          if (!forOpening && settings.get_boolean('broken-glass-use-pointer')) {
+            const [x, y]               = global.get_pointer();
+            const [ok, localX, localY] = actor.transform_stage_point(x, y);
 
-      // This is required to bind the dust texture for drawing. Sadly, this seems to be
+            if (ok) {
+              epicenterX = localX / actor.width;
+              epicenterY = localY / actor.height;
+            }
+          }
+
+          // clang-format off
+          shader.set_uniform_float(shader._uSeed,       2, [testMode ? 0 : Math.random(), testMode ? 0 : Math.random()]);
+          shader.set_uniform_float(shader._uEpicenter,  2, [epicenterX, epicenterY]);
+          shader.set_uniform_float(shader._uShardScale, 1, [settings.getValue('broken-glass-scale')]);
+          shader.set_uniform_float(shader._uBlowForce,  1, [settings.getValue('broken-glass-blow-force')]);
+          shader.set_uniform_float(shader._uGravity,    1, [settings.getValue('broken-glass-gravity')]);
+          // clang-format on
+        });
+
+      // This is required to bind the shard texture for drawing. Sadly, this seems to be
       // impossible under GNOME 3.3x as this.get_pipeline() is not available. It was
       // called get_target() back then but this is not wrapped in GJS.
       // https://gitlab.gnome.org/GNOME/mutter/-/blob/gnome-3-36/clutter/clutter/clutter-offscreen-effect.c#L598
@@ -95,10 +122,10 @@ var Effect = class Effect {
         pipeline.set_layer_filters(0, Cogl.PipelineFilter.LINEAR,
                                    Cogl.PipelineFilter.LINEAR);
 
-        // Bind the dust texture.
-        pipeline.set_layer_texture(1, this._dustTexture.get_texture());
+        // Bind the shard texture.
+        pipeline.set_layer_texture(1, this._shardTexture.get_texture());
         pipeline.set_layer_wrap_mode(1, Cogl.PipelineWrapMode.REPEAT);
-        pipeline.set_uniform_1i(shader._uDustTexture, 1);
+        pipeline.set_uniform_1i(shader._uShardTexture, 1);
       });
     });
   }
@@ -113,15 +140,16 @@ var Effect = class Effect {
   // This will be called in various places where a unique identifier for this effect is
   // required. It should match the prefix of the settings keys which store whether the
   // effect is enabled currently (e.g. '*-enable-effect'), and its animation time
-  // (e.g. '*-animation-time').
+  // (e.g. '*-animation-time'). Also, the shader file and the settings UI files should be
+  // named likes this.
   static getNick() {
-    return 'snap';
+    return 'broken-glass';
   }
 
   // This will be shown in the sidebar of the preferences dialog as well as in the
   // drop-down menus where the user can choose the effect.
   static getLabel() {
-    return _('Snap of Disintegration');
+    return _('Broken Glass');
   }
 
   // -------------------------------------------------------------------- API for prefs.js
@@ -129,9 +157,11 @@ var Effect = class Effect {
   // This is called by the preferences dialog whenever a new effect profile is loaded. It
   // binds all user interface elements to the respective settings keys of the profile.
   static bindPreferences(dialog) {
-    dialog.bindAdjustment('snap-animation-time');
-    dialog.bindAdjustment('snap-scale');
-    dialog.bindColorButton('snap-color');
+    dialog.bindAdjustment('broken-glass-animation-time');
+    dialog.bindAdjustment('broken-glass-scale');
+    dialog.bindAdjustment('broken-glass-gravity');
+    dialog.bindAdjustment('broken-glass-blow-force');
+    dialog.bindSwitch('broken-glass-use-pointer');
   }
 
   // ---------------------------------------------------------------- API for extension.js
@@ -140,6 +170,6 @@ var Effect = class Effect {
   // animation. This is useful if the effect requires drawing something beyond the usual
   // bounds of the actor. This only works for GNOME 3.38+.
   static getActorScale(settings, forOpening, actor) {
-    return {x: 1.2, y: 1.2};
+    return {x: 2.0, y: 2.0};
   }
 }
