+++ CinnamonBurnMyWindows@klangman/files/CinnamonBurnMyWindows@klangman/effects/EnergizeA.js
