+++ CinnamonBurnMyWindows@klangman/files/CinnamonBurnMyWindows@klangman/effects/PixelWipe.js
